Implementando o Spring Security

+++ app/src/main/java/application/model/Login.java
@@ -8,41 +8,34 @@ import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 
 @Entity
-@Table(name="login")
+@Table(name = "login")
 public class Login {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
+    private int id;
     @Column(nullable = false)
     private String email;
-
     @Column(nullable = false)
     private String senha;
 
-    public long getId() {
+    public int getId() {
         return id;
     }
-
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
-
     public String getEmail() {
         return email;
     }
-
     public void setEmail(String email) {
         this.email = email;
     }
-
     public String getSenha() {
         return senha;
     }
-
     public void setSenha(String senha) {
         this.senha = senha;
     }
-    
+
     
 }
